backend/requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
catboost==1.2.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cmdstanpy==1.3.0
colorlog==6.10.1
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
holidays==0.83
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lightgbm==4.6.0
llvmlite==0.43.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
narwhals==2.16.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
optuna==4.8.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.6.0
prophet==1.3.0
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
resend==2.23.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.1.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.46
stanio==0.5.1
starlette==0.49.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.1
torch==2.8.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
xgboost==2.1.4
zipp==3.23.0
